* Switched all object types to use uuids as their keys, instead of sequential ids. That is, I made the change then got all the compile-time errors fixed; I still need to fix any runtime errors that come up.
Why switching to uuids? Because it greatly reduces (basically to 0) the chance of ID conflicts when two users are working concurrently, and removes the need for a "lastID" flag in the database. (you can solve id-conflicts with sequential ids by various means, but for client-centric apps like this, it's very nice to be able to just generate everything locally and have it virtually always succeed with a direct firestore add-db-entry call)