* Started transition to MobX: Moved one store value from Redux into the new MobX store. (playingTimeline_time) * Fixed some potential bugs with null entries in GetData("myCollection") calls. (discovered through errors in test)
Note: Performance is (substantially) lower now since the dependency-checking of PathWatchManager is disabled (so as to not conflict with MobX's dependency-tracking).
=== Tests:
Test failed to complete, because it took to long to even load the map ui. It appears to be due to each Link ui *actually applying* its actions, just from existing on the page. (Links are only supposed to pre-calculate their effect on the redux store's state, not actually apply it; I guess I broke it earlier with part of my work on PathWatchManager -- else more recently with starting MobX integration, though I don't see how)